+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - [#2413] AS3 direct editation - try/catch clause in instance initializer
 - [#2386] Editor mode - not able to save shape/morphshape points
 - Loading icon did not respect UI scale factor
+- Editor (JSyntaxPane) incorrectly draws line numbers panel on Hi-dpi displays
 
 ## [22.0.2] - 2025-01-17
 ### Added